Use caching in Interpreter
